refactor(gateway): GAR-480 — Q10.f extract build_imessage_channels to bootstrap/imessage.rs
Moves the macOS-only `build_imessage_channels` (~123 LOC) out of
`bootstrap/mod.rs` into a new sibling `bootstrap/imessage.rs`, following
the extract-and-re-export pattern established by slices 10.b–10.e.

- New file: `bootstrap/imessage.rs` (macOS-gated via `#![cfg(target_os = "macos")]`)
- `bootstrap/mod.rs`: removes inline function + cfg-gated IMessage imports;
  adds `#[cfg(target_os = "macos")] mod imessage;` + `pub use imessage::build_imessage_channels;`
- Zero behaviour change; all call sites in `server.rs` unchanged
- `bootstrap/mod.rs` shrinks ~1699 → ~1576 LOC (−123)

Plan: plans/0168-gar-480-q10f-bootstrap-imessage.md
Linear: GAR-480
