+++ plans/README.md
@@ -177,3 +177,4 @@ Este diretório contém os planos de implementação para o projeto GarraRUST.
 | 0165 | [GAR-477 — Q10.c: extract `bootstrap/discord.rs` (270 LOC, zero behaviour change)](0165-gar-477-q10c-bootstrap-discord.md) | [GAR-477](https://linear.app/chatgpt25/issue/GAR-477) | ✅ Merged 2026-05-22 via PR #471 (`b594ace`) |
 | 0166 | [GAR-478 — Q10.d: extract `bootstrap/slack.rs` (166 LOC, zero behaviour change)](0166-gar-478-q10d-bootstrap-slack.md) | [GAR-478](https://linear.app/chatgpt25/issue/GAR-478) | ✅ Merged 2026-05-22 via PR #474 (`4a51841`) |
 | 0167 | [GAR-479 — Q10.e: extract `bootstrap/whatsapp.rs` (180 LOC, zero behaviour change)](0167-gar-479-q10e-bootstrap-whatsapp.md) | [GAR-479](https://linear.app/chatgpt25/issue/GAR-479) | ✅ Merged 2026-05-22 via PR #476 (`60a8dff`) |
+| 0168 | [GAR-480 — Q10.f: extract `bootstrap/imessage.rs` (123 LOC, macOS-only, zero behaviour change)](0168-gar-480-q10f-bootstrap-imessage.md) | [GAR-480](https://linear.app/chatgpt25/issue/GAR-480) | 🔄 In Progress |
